requirements.txt
pydantic>=2.10.0,<3.0.0
PyYAML>=6.0,<7.0
docker>=7.0.0,<8.0.0
click>=8.0.0,<9.0.0 